fix min and max having redefined errors
`min` and `max` were having redefined errors when WinAPI was also included. Min and Max are very common terms and it is reasonable to expect other libraries that may be included alongside sr_resolve to also have macros, functions, etc. to be named min and max. Renaming it SR_RESOLVE_MAX eliminates this common issue